src/youtube_qa/vector_search.py
"""Vector search management for YouTube transcript chunks."""

from databricks.sdk import WorkspaceClient
from databricks.vector_search import VectorSearchIndex
from databricks.vector_search.client import VectorSearchClient
from loguru import logger

from youtube_qa.config import ProjectConfig


class VectorSearchManager:
    """Manages vector search endpoints and indexes for YouTube transcript chunks."""

    def __init__(
        self,
        config: ProjectConfig,
        endpoint_name: str | None = None,
        embedding_model: str | None = None,
        usage_policy_id: str | None = None,
    ) -> None:
        """Initialize VectorSearchManager."""
        self.config = config
        self.endpoint_name = endpoint_name or config.vector_search_endpoint
        self.embedding_model = embedding_model or config.embedding_endpoint
        self.catalog = config.catalog
        self.schema = config.schema
        self.usage_policy_id = usage_policy_id

        w = WorkspaceClient()
        self.client = VectorSearchClient(
            workspace_url=w.config.host,
            personal_access_token=w.tokens.create(lifetime_seconds=1200).token_value,
        )

        self.index_name = f"{self.catalog}.{self.schema}.youtube_index"

    def create_endpoint_if_not_exists(self) -> None:
        """Create vector search endpoint if it doesn't exist."""
        endpoints_response = self.client.list_endpoints()
        endpoints = (
            endpoints_response.get("endpoints", []) if isinstance(endpoints_response, dict) else []
        )

        endpoint_exists = any(
            (ep.get("name") if isinstance(ep, dict) else getattr(ep, "name", None))
            == self.endpoint_name
            for ep in endpoints
        )

        if endpoint_exists:
            logger.info(f"✓ Vector search endpoint exists: {self.endpoint_name}")
            return

        logger.info(f"Creating vector search endpoint: {self.endpoint_name}")
        self.client.create_endpoint_and_wait(
            name=self.endpoint_name,
            endpoint_type="STANDARD",
            usage_policy_id=self.usage_policy_id,
        )
        logger.info(f"✓ Vector search endpoint created: {self.endpoint_name}")

    def create_or_get_index(self) -> VectorSearchIndex:
        """Create or get the vector search index."""
        self.create_endpoint_if_not_exists()
        source_table = f"{self.catalog}.{self.schema}.youtube_chunks_table"

        try:
            index = self.client.get_index(index_name=self.index_name)
            logger.info(f"✓ Vector search index exists: {self.index_name}")
            return index
        except Exception:
            logger.info(f"Index {self.index_name} not found, will create it")

        try:
            index = self.client.create_delta_sync_index(
                endpoint_name=self.endpoint_name,
                source_table_name=source_table,
                index_name=self.index_name,
                pipeline_type="TRIGGERED",
                primary_key="id",
                embedding_source_column="text",
                embedding_model_endpoint_name=self.embedding_model,
                usage_policy_id=self.usage_policy_id,
            )
            logger.info(f"✓ Vector search index created: {self.index_name}")
            return index
        except Exception as e:
            if "RESOURCE_ALREADY_EXISTS" not in str(e):
                raise

            logger.info(f"✓ Vector search index exists: {self.index_name}")
            return self.client.get_index(index_name=self.index_name)

    def sync_index(self) -> None:
        """Sync the vector search index with its source Delta table."""
        index = self.create_or_get_index()
        logger.info(f"Syncing vector search index: {self.index_name}")
        index.sync()
        logger.info("✓ Index sync triggered")

    def search(self, query: str, num_results: int = 5, filters: dict | None = None) -> dict:
        """Run a similarity search query."""
        index = self.client.get_index(index_name=self.index_name)
        return index.similarity_search(
            query_text=query,
            columns=["id", "text", "metadata"],
            num_results=num_results,
            filters=filters,
        )
